pyproject.toml
[tool.poetry]
name = "bumblebee_backend"
version = "0.2.0"
description = "A social media for YOU"
authors = ["The Godfathers"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework-simplejwt = "^4.6.0"
djangorestframework = "^3.12.4"
psycopg2 = "^2.8.6"
Pillow = "^8.2.0"
python-dotenv = "^0.17.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
isort = "^5.8.0"
black = "^21.4b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
